requirements.txt
fastapi
uvicorn
torch
torchvision
pandas
Pillow
requests
